Refactoring inotify code to a base interface
